Move before_request hook into provider
